pointcloud/m42611: change geometry representation from double to int32_t
Since the codec can only code integer positions, using Vec3<double> for
position data requires repeated inefficient format conversions.  This
commit changes the internal representation to Vec3<int32_t>.

+++ tmc3/hls.h
@@ -356,7 +356,7 @@ struct AttributeParameterSet {
   int max_num_direct_predictors;
   int adaptive_prediction_threshold;
   int search_range;
-  Vec3<double> lod_neigh_bias;
+  Vec3<int32_t> lod_neigh_bias;
   bool intra_lod_prediction_enabled_flag;
   bool inter_component_prediction_enabled_flag;
